norm#4 (but NOT IN NORM)

+++ key_left_or_right.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   key_left_right.c                                   :+:      :+:    :+:   */
+/*   key_left_or_right.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: dwinky <dwinky@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/24 21:00:05 by dwinky            #+#    #+#             */
-/*   Updated: 2021/04/26 20:48:35 by dwinky           ###   ########.fr       */
+/*   Updated: 2021/04/27 18:06:35 by dwinky           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 			/** repeat command quantity times */
 static void	repeat_command(char *command, int quantity)
 {
-	int k;
+	int	k;
 
 	k = 0;
 	while (k < quantity)
@@ -27,9 +27,9 @@ static void	repeat_command(char *command, int quantity)
 
 void	key_left_or_right(int *cursor_pos, char *str, char *line)
 {
-	int width_term;
+	int	width_term;
 
-	width_term = tgetnum("columns"); // define in term.h
+	width_term = tgetnum("columns");
 	if (!strcmp(str, "\e[D"))
 	{
 		if ((*cursor_pos + WIDTH_PROMT) >= width_term && (*cursor_pos + WIDTH_PROMT) % width_term == 0)
@@ -46,13 +46,13 @@ void	key_left_or_right(int *cursor_pos, char *str, char *line)
 	}
 	else if (!strcmp(str, "\e[C"))
 	{
-		if (*cursor_pos < ft_strlen(line) && (*cursor_pos + WIDTH_PROMT - 1) == width_term)
+		if ((size_t)(*cursor_pos) < ft_strlen(line) && (*cursor_pos + WIDTH_PROMT - 1) == width_term)
 		{
 			tputs(cursor_down, 1, ft_putchar);
 			repeat_command(cursor_left, width_term);
 			(*cursor_pos)++;
 		}
-		else if (0 <= *cursor_pos && *cursor_pos < ft_strlen(line))
+		else if (0 <= *cursor_pos && (size_t)(*cursor_pos) < ft_strlen(line))
 		{
 			tputs(cursor_right, 1, ft_putchar);
 			(*cursor_pos)++;
